+++ .github/scripts/test_downstream_project.sh
@@ -1,0 +1,7 @@
+pushd ../$PROJECT
+python -m pytest -vl
+TEST_RETURN_CODE=$?
+popd
+if [[ "$TEST_RETURN_CODE" != "0" ]]; then
+    exit $TEST_RETURN_CODE
+fi